Rearrange Seek Methods
