+++ src/System.Buffers.Primitives/System/Buffers/ReadOnlyBuffer_helpers.cs
@@ -10,8 +10,65 @@ namespace System.Buffers
 {
     public readonly partial struct ReadOnlyBuffer 
     {
+        public static int Seek(Position begin, Position end, out Position result, byte byte0)
+        {
+            var enumerator = new ReadOnlyBuffer(begin, end).GetEnumerator();
+            while (enumerator.MoveNext())
+            {
+                var span = enumerator.Current.Span;
+
+                int index = span.IndexOf(byte0);
+                if (index != -1)
+                {
+                    result = enumerator.CreateCursor(index);
+                    return span[index];
+                }
+            }
+
+            result = end;
+            return -1;
+        }
+
+        public static int Seek(Position begin, Position end, out Position result, byte byte0, byte byte1)
+        {
+            var enumerator = new ReadOnlyBuffer(begin, end).GetEnumerator();
+            while (enumerator.MoveNext())
+            {
+                var span = enumerator.Current.Span;
+                int index = span.IndexOfAny(byte0, byte1);
+
+                if (index != -1)
+                {
+                    result = enumerator.CreateCursor(index);
+                    return span[index];
+                }
+            }
+
+            result = end;
+            return -1;
+        }
+
+        public static int Seek(Position begin, Position end, out Position result, byte byte0, byte byte1, byte byte2)
+        {
+            var enumerator = new ReadOnlyBuffer(begin, end).GetEnumerator();
+            while (enumerator.MoveNext())
+            {
+                var span = enumerator.Current.Span;
+                int index = span.IndexOfAny(byte0, byte1, byte2);
+
+                if (index != -1)
+                {
+                    result = enumerator.CreateCursor(index);
+                    return span[index];
+                }
+            }
+
+            result = end;
+            return -1;
+        }
+
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static bool TryGetBuffer(Position begin, Position end, out ReadOnlyMemory<byte> data, out Position next)
+        internal static bool TryGetBuffer(Position begin, Position end, out ReadOnlyMemory<byte> data, out Position next)
         {
             var segment = begin.Segment;
 
@@ -82,7 +139,7 @@ namespace System.Buffers
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        private static Position Seek(Position begin, Position end, long bytes, bool checkEndReachable = true)
+        internal static Position Seek(Position begin, Position end, long bytes, bool checkEndReachable = true)
         {
             Position cursor;
             if (begin.Segment == end.Segment && end.Index - begin.Index >= bytes)
