+++ tests/System.IO.Pipelines.Tests/SeekTests.cs
@@ -77,15 +77,15 @@ namespace System.IO.Pipelines.Tests
             int found = -1;
             if (searchFor.Length == 1)
             {
-                found = PositionOperations.Seek(start, end, out result, (byte)searchFor[0]);
+                found = ReadOnlyBuffer.Seek(start, end, out result, (byte)searchFor[0]);
             }
             else if (searchFor.Length == 2)
             {
-                found = PositionOperations.Seek(start, end, out result, (byte)searchFor[0], (byte)searchFor[1]);
+                found = ReadOnlyBuffer.Seek(start, end, out result, (byte)searchFor[0], (byte)searchFor[1]);
             }
             else if (searchFor.Length == 3)
             {
-                found = PositionOperations.Seek(start, end, out result, (byte)searchFor[0], (byte)searchFor[1], (byte)searchFor[2]);
+                found = ReadOnlyBuffer.Seek(start, end, out result, (byte)searchFor[0], (byte)searchFor[1], (byte)searchFor[2]);
             }
             else
             {
@@ -105,9 +105,9 @@ namespace System.IO.Pipelines.Tests
 
             // Act
             var end = limit > input.Length ? buffer.End : buffer.Slice(0, limit).End;
-            var returnValue = PositionOperations.Seek(buffer.Start, end, out Position result, (byte)seek);
-            var returnValue_1 = PositionOperations.Seek(buffer.Start, end, out result, (byte)seek, (byte)seek);
-            var returnValue_2 = PositionOperations.Seek(buffer.Start, end, out result, (byte)seek, (byte)seek, (byte)seek);
+            var returnValue = ReadOnlyBuffer.Seek(buffer.Start, end, out Position result, (byte)seek);
+            var returnValue_1 = ReadOnlyBuffer.Seek(buffer.Start, end, out result, (byte)seek, (byte)seek);
+            var returnValue_2 = ReadOnlyBuffer.Seek(buffer.Start, end, out result, (byte)seek, (byte)seek, (byte)seek);
 
             // Assert
             Assert.Equal(expectedReturnValue, returnValue);
@@ -141,17 +141,17 @@ namespace System.IO.Pipelines.Tests
             var end = buffer.End;
 
             // Act
-            var endReturnValue = PositionOperations.Seek(start, veryEnd, out end, (byte)limitAfter);
+            var endReturnValue = ReadOnlyBuffer.Seek(start, veryEnd, out end, (byte)limitAfter);
             if (endReturnValue != -1)
             {
                 end = buffer.Slice(end, 1).End;
             }
-            var returnValue1 = PositionOperations.Seek(start, end, out scan1, (byte)seek);
-            var returnValue2_1 = PositionOperations.Seek(start, end, out scan2_1, (byte)seek, afterSeek);
-            var returnValue2_2 = PositionOperations.Seek(start, end, out scan2_2, afterSeek, (byte)seek);
-            var returnValue3_1 = PositionOperations.Seek(start, end, out scan3_1, (byte)seek, afterSeek, afterSeek);
-            var returnValue3_2 = PositionOperations.Seek(start, end, out scan3_2, afterSeek, (byte)seek, afterSeek);
-            var returnValue3_3 = PositionOperations.Seek(start, end, out scan3_3, afterSeek, afterSeek, (byte)seek);
+            var returnValue1 = ReadOnlyBuffer.Seek(start, end, out scan1, (byte)seek);
+            var returnValue2_1 = ReadOnlyBuffer.Seek(start, end, out scan2_1, (byte)seek, afterSeek);
+            var returnValue2_2 = ReadOnlyBuffer.Seek(start, end, out scan2_2, afterSeek, (byte)seek);
+            var returnValue3_1 = ReadOnlyBuffer.Seek(start, end, out scan3_1, (byte)seek, afterSeek, afterSeek);
+            var returnValue3_2 = ReadOnlyBuffer.Seek(start, end, out scan3_2, afterSeek, (byte)seek, afterSeek);
+            var returnValue3_3 = ReadOnlyBuffer.Seek(start, end, out scan3_3, afterSeek, afterSeek, (byte)seek);
 
 
             // Assert
